Add files via upload
As a user, I want to receive two things when they create a request: to be able to track the status of their supply request. This means that the user wants to have access to information about the progress of their request, including whether it has been fulfilled and location of the cabinet where they can pick up their order. The user wants this feature so that they can have a record of their order and know where to go to pick it up. Also the app can help the user keep track of their orders and make the process of picking them up more efficient and convenient. This user story can help guide the development of the app feature, making sure it meets the user's needs and expectations.